refactor: rename Calendar folder to CalendarLayout
